Relocate config_test to tests package and Export struct members of GatewayConfig{}
