+++ gateway/tests/config_test.go
@@ -1,11 +1,13 @@
 // Copyright (c) Alex Ellis 2017. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-package main
+package tests
 
 import (
 	"testing"
 	"time"
+
+	gateway "github.com/alexellis/faas/gateway"
 )
 
 type EnvBucket struct {
@@ -28,16 +30,16 @@ func (e EnvBucket) Setenv(key string, value string) {
 
 func TestRead_EmptyTimeoutConfig(t *testing.T) {
 	defaults := NewEnvBucket()
-	readConfig := ReadConfig{}
+	readConfig := gateway.ReadConfig{}
 
 	config := readConfig.Read(defaults)
 
-	if (config.readTimeout) != time.Duration(8)*time.Second {
-		t.Log("readTimeout incorrect")
+	if (config.ReadTimeout) != time.Duration(8)*time.Second {
+		t.Log("ReadTimeout incorrect")
 		t.Fail()
 	}
-	if (config.writeTimeout) != time.Duration(8)*time.Second {
-		t.Log("writeTimeout incorrect")
+	if (config.WriteTimeout) != time.Duration(8)*time.Second {
+		t.Log("WriteTimeout incorrect")
 		t.Fail()
 	}
 }
@@ -47,15 +49,15 @@ func TestRead_ReadAndWriteTimeoutConfig(t *testing.T) {
 	defaults.Setenv("read_timeout", "10")
 	defaults.Setenv("write_timeout", "60")
 
-	readConfig := ReadConfig{}
+	readConfig := gateway.ReadConfig{}
 	config := readConfig.Read(defaults)
 
-	if (config.readTimeout) != time.Duration(10)*time.Second {
-		t.Logf("readTimeout incorrect, got: %d\n", config.readTimeout)
+	if (config.ReadTimeout) != time.Duration(10)*time.Second {
+		t.Logf("ReadTimeout incorrect, got: %d\n", config.ReadTimeout)
 		t.Fail()
 	}
-	if (config.writeTimeout) != time.Duration(60)*time.Second {
-		t.Logf("writeTimeout incorrect, got: %d\n", config.writeTimeout)
+	if (config.WriteTimeout) != time.Duration(60)*time.Second {
+		t.Logf("WriteTimeout incorrect, got: %d\n", config.WriteTimeout)
 		t.Fail()
 	}
 }
